parsers: factor out low-level routines to utils
