+++ aiida_cp2k/parsers.py
@@ -7,19 +7,15 @@
 ###############################################################################
 """AiiDA-CP2K output parser"""
 from __future__ import absolute_import
-from __future__ import division
 
 import io
 import os
-import re
-import math
-
-import ase
-import numpy as np
 
 from aiida.parsers import Parser
 from aiida.common import OutputParsingError, NotExistent
 from aiida.engine import ExitCode
+
+from .utils import parse_cp2k_output, parse_cp2k_trajectory
 
 
 class Cp2kParser(Parser):
@@ -59,83 +55,28 @@ class Cp2kParser(Parser):
         abs_fn = os.path.join(out_folder._repository._get_base_folder().abspath, fname)
 
         with io.open(abs_fn, mode="r", encoding="utf-8") as fobj:
-            lines = fobj.readlines()
-
-        result_dict = {"exceeded_walltime": False}
-
-        for i_line, line in enumerate(lines):
-            if line.startswith(" ENERGY| "):
-                result_dict["energy"] = float(line.split()[8])
-                result_dict["energy_units"] = "a.u."
-            elif "The number of warnings for this run is" in line:
-                result_dict["nwarnings"] = int(line.split()[-1])
-            elif "exceeded requested execution time" in line:
-                result_dict["exceeded_walltime"] = True
-            elif "KPOINTS| Band Structure Calculation" in line:
-                kpoints, labels, bands = self._parse_bands(lines, i_line)
-                bnds = BandsData()
-                bnds.set_kpoints(kpoints)
-                bnds.labels = labels
-                bnds.set_bands(bands, units="eV")
-                self.out("output_bands", bnds)
-            else:
-                # ignore all other lines
-                pass
+            result_dict = parse_cp2k_output(fobj)
 
         if "nwarnings" not in result_dict:
             raise OutputParsingError("CP2K did not finish properly.")
 
+        if "kpoint_data" in result_dict:
+            bnds = BandsData()
+            bnds.set_kpoints(result_dict["kpoint_data"]["kpoints"])
+            bnds.labels = result_dict["kpoint_data"]["labels"]
+            bnds.set_bands(
+                result_dict["kpoint_data"]["bands"],
+                units=result_dict["kpoint_data"]["bands_unit"],
+            )
+            self.out("output_bands", bnds)
+            del result_dict["kpoint_data"]
+
         self.out("output_parameters", Dict(dict=result_dict))
-
-    @staticmethod
-    def _parse_bands(lines, n_start):
-        """Parse band structure from cp2k output"""
-        kpoints = []
-        labels = []
-        bands_s1 = []
-        bands_s2 = []
-        known_kpoints = {}
-        pattern = re.compile(".*?Nr.*?Spin.*?K-Point.*?", re.DOTALL)
-
-        selected_lines = lines[n_start:]
-        for current_line, line in enumerate(selected_lines):
-            splitted = line.split()
-            if "KPOINTS| Special K-Point" in line:
-                kpoint = tuple(float(p) for p in splitted[-3:])
-                if " ".join(splitted[-5:-3]) != "not specified":
-                    label = splitted[-4]
-                    known_kpoints[kpoint] = label
-            elif pattern.match(line):
-                spin = int(splitted[3])
-                kpoint = tuple(float(p) for p in splitted[-3:])
-                kpoint_n_lines = int(
-                    math.ceil(int(selected_lines[current_line + 1]) / 4)
-                )
-                band = [
-                    float(v)
-                    for v in " ".join(
-                        selected_lines[
-                            current_line + 2 : current_line + 2 + kpoint_n_lines
-                        ]
-                    ).split()
-                ]
-                if spin == 1:
-                    if kpoint in known_kpoints:
-                        labels.append((len(kpoints), known_kpoints[kpoint]))
-                    kpoints.append(kpoint)
-                    bands_s1.append(band)
-                elif spin == 2:
-                    bands_s2.append(band)
-        if bands_s2:
-            bands = [bands_s1, bands_s2]
-        else:
-            bands = bands_s1
-
-        return np.array(kpoints), labels, np.array(bands)
 
     def _parse_trajectory(self, out_folder):
         """CP2K trajectory parser"""
 
+        from ase import Atoms
         from aiida.orm import StructureData
 
         # pylint: disable=protected-access
@@ -151,22 +92,6 @@ class Cp2kParser(Parser):
         abs_fn = os.path.join(out_folder._repository._get_base_folder().abspath, fname)
 
         with io.open(abs_fn, mode="r", encoding="utf-8") as fobj:
-            content = fobj.read()
-
-        # parse coordinate section
-        match = re.search(r"\n\s*&COORD\n(.*?)\n\s*&END COORD\n", content, re.DOTALL)
-        coord_lines = [line.strip().split() for line in match.group(1).splitlines()]
-        symbols = [line[0] for line in coord_lines]
-        positions_str = [line[1:] for line in coord_lines]
-        positions = np.array(positions_str, np.float64)
-
-        # parse cell section
-        match = re.search(r"\n\s*&CELL\n(.*?)\n\s*&END CELL\n", content, re.DOTALL)
-        cell_lines = [line.strip().split() for line in match.group(1).splitlines()]
-        cell_str = [line[1:] for line in cell_lines if line[0] in "ABC"]
-        cell = np.array(cell_str, np.float64)
-
-        # create StructureData
-        atoms = ase.Atoms(symbols=symbols, positions=positions, cell=cell)
+            atoms = Atoms(**parse_cp2k_trajectory(fobj))
 
         return StructureData(ase=atoms)
