+++ aiida_cp2k/utils.py
@@ -8,8 +8,12 @@
 """AiiDA-CP2K input plugin"""
 
 from __future__ import absolute_import
+from __future__ import division
+
 from itertools import chain
 from copy import deepcopy
+import re
+import math
 
 import six
 
@@ -144,3 +148,103 @@ class Cp2kInput:
 
             else:
                 yield "{ispace}{key} {val}".format(ispace=ispace, key=key, val=val)
+
+
+def parse_cp2k_output(fobj):
+    lines = fobj.readlines()
+
+    result_dict = {"exceeded_walltime": False}
+
+    for i_line, line in enumerate(lines):
+        if line.startswith(" ENERGY| "):
+            result_dict["energy"] = float(line.split()[8])
+            result_dict["energy_units"] = "a.u."
+        elif "The number of warnings for this run is" in line:
+            result_dict["nwarnings"] = int(line.split()[-1])
+        elif "exceeded requested execution time" in line:
+            result_dict["exceeded_walltime"] = True
+        elif "KPOINTS| Band Structure Calculation" in line:
+            kpoints, labels, bands = _parse_bands(lines, i_line)
+            result_dict["kpoint_data"] = {
+                "kpoints": kpoints,
+                "labels": labels,
+                "bands": bands,
+                "bands_unit": "eV",
+            }
+        else:
+            # ignore all other lines
+            pass
+
+    return result_dict
+
+
+def _parse_bands(lines, n_start):
+    """Parse band structure from cp2k output"""
+
+    import numpy as np
+
+    kpoints = []
+    labels = []
+    bands_s1 = []
+    bands_s2 = []
+    known_kpoints = {}
+    pattern = re.compile(".*?Nr.*?Spin.*?K-Point.*?", re.DOTALL)
+
+    selected_lines = lines[n_start:]
+    for current_line, line in enumerate(selected_lines):
+        splitted = line.split()
+        if "KPOINTS| Special K-Point" in line:
+            kpoint = tuple(float(p) for p in splitted[-3:])
+            if " ".join(splitted[-5:-3]) != "not specified":
+                label = splitted[-4]
+                known_kpoints[kpoint] = label
+        elif pattern.match(line):
+            spin = int(splitted[3])
+            kpoint = tuple(float(p) for p in splitted[-3:])
+            kpoint_n_lines = int(math.ceil(int(selected_lines[current_line + 1]) / 4))
+            band = [
+                float(v)
+                for v in " ".join(
+                    selected_lines[current_line + 2 : current_line + 2 + kpoint_n_lines]
+                ).split()
+            ]
+
+            if spin == 1:
+                if kpoint in known_kpoints:
+                    labels.append((len(kpoints), known_kpoints[kpoint]))
+                kpoints.append(kpoint)
+                bands_s1.append(band)
+            elif spin == 2:
+                bands_s2.append(band)
+
+    if bands_s2:
+        bands = [bands_s1, bands_s2]
+    else:
+        bands = bands_s1
+
+    return np.array(kpoints), labels, np.array(bands)
+
+
+def parse_cp2k_trajectory(fobj):
+    """CP2K trajectory parser"""
+
+    import numpy as np
+
+    # pylint: disable=protected-access
+
+    content = fobj.read()
+
+    # parse coordinate section
+    match = re.search(r"\n\s*&COORD\n(.*?)\n\s*&END COORD\n", content, re.DOTALL)
+    coord_lines = [line.strip().split() for line in match.group(1).splitlines()]
+    symbols = [line[0] for line in coord_lines]
+    positions_str = [line[1:] for line in coord_lines]
+    positions = np.array(positions_str, np.float64)
+
+    # parse cell section
+    match = re.search(r"\n\s*&CELL\n(.*?)\n\s*&END CELL\n", content, re.DOTALL)
+    cell_lines = [line.strip().split() for line in match.group(1).splitlines()]
+    cell_str = [line[1:] for line in cell_lines if line[0] in "ABC"]
+    cell = np.array(cell_str, np.float64)
+
+    return {"symbols": symbols, "positions": positions, "cell": cell}
